更新group功能

+++ Src/AlonsoAdmin.Services/System/Implement/SysGroupService.cs
@@ -8,6 +8,7 @@ using AlonsoAdmin.Services.System.Response;
 using AutoMapper;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -25,6 +26,7 @@ namespace AlonsoAdmin.Services.System.Implement
             _mapper = mapper;
             _sysGroupRepository = sysGroupRepository;
         }
+
         public async Task<IResponseEntity> CreateAsync(GroupAddRequest req)
         {
             var item = _mapper.Map<SysGroupEntity>(req);
@@ -32,29 +34,73 @@ namespace AlonsoAdmin.Services.System.Implement
             return ResponseEntity.Result(result != null && result?.Id != "");
         }
 
+        public async Task<IResponseEntity> UpdateAsync(GroupEditRequest req)
+        {
+
+            if (req == null || req?.Id == "")
+            {
+                return ResponseEntity.Error("更新的实体主键丢失");
+            }
+
+            //var entity = await _sysGroupRepository.GetAsync(req.Id);
+            //if (entity == null || entity?.Id == "")
+            //{
+            //    return ResponseEntity.Error("找不到更新的实体！");
+            //}
+            //_mapper.Map(req, entity);
+
+            var entity = _mapper.Map<SysGroupEntity>(req);
+            await _sysGroupRepository.UpdateAsync(entity);
+            return ResponseEntity.Ok("更新成功");
+        }
+
         public async Task<IResponseEntity> DeleteAsync(string id)
         {
+            if (id == null || id == "")
+            {
+                return ResponseEntity.Error("删除对象的主键获取失败");
+            }
             var result = await _sysGroupRepository.DeleteAsync(id);
             return ResponseEntity.Result(result > 0);
         }
 
-        public Task<IResponseEntity> DeleteBatchAsync(string[] ids)
+        public async Task<IResponseEntity> DeleteBatchAsync(string[] ids)
         {
-            throw new NotImplementedException();
+            if (ids == null || ids.Length == 0)
+            {
+                return ResponseEntity.Error("删除对象的主键获取失败");
+            }
+            var result = await _sysGroupRepository.Where(m => ids.Contains(m.Id)).ToDelete().ExecuteAffrowsAsync();
+            return ResponseEntity.Result(result > 0);
         }
 
-        public async Task<IResponseEntity> GetAllAsync(GroupFilterRequest req)
+        public async Task<IResponseEntity> SoftDeleteAsync(string id)
         {
-            var key = req?.Key;
+            if (id == null || id == "")
+            {
+                return ResponseEntity.Error("删除对象的主键获取失败");
+            }
 
-            var list = await _sysGroupRepository.Select
-                .WhereIf(key.IsNotNull(), a => (a.Title.Contains(key) || a.Code.Contains(key)))
-                .Count(out var total)
-                .OrderBy(true, a => a.OrderIndex)
-                .ToListAsync();
+            var result = await _sysGroupRepository.SoftDeleteAsync(id);
+            return ResponseEntity.Result(result);
+        }
 
-            var result = _mapper.Map<List<GroupListResponse>>(list);
-            return ResponseEntity.Ok(result);
+        public async Task<IResponseEntity> SoftDeleteBatchAsync(string[] ids)
+        {
+            if (ids == null || ids.Length == 0)
+            {
+                return ResponseEntity.Error("删除对象的主键获取失败");
+            }
+
+            var result = await _sysGroupRepository.SoftDeleteAsync(ids);
+            return ResponseEntity.Result(result);
+        }
+
+        public async Task<IResponseEntity> GetItemAsync(string id)
+        {
+
+            var result = await _sysGroupRepository.GetAsync(id);
+            return ResponseEntity.Result(result != null);
         }
 
         public async Task<IResponseEntity> GetListAsync(RequestEntity<GroupFilterRequest> req)
@@ -77,40 +123,22 @@ namespace AlonsoAdmin.Services.System.Implement
             return ResponseEntity.Ok(data);
         }
 
-        public async Task<IResponseEntity> SoftDeleteAsync(string id)
+        public async Task<IResponseEntity> GetAllAsync(GroupFilterRequest req)
         {
-            if (id == null || id == "")
-            {
-                return ResponseEntity.Error("删除对象的主键获取失败");
-            }
+            var key = req?.Key;
 
-            var result = await _sysGroupRepository.SoftDeleteAsync(id);
-            return ResponseEntity.Result(result );
+            var list = await _sysGroupRepository.Select
+                .WhereIf(key.IsNotNull(), a => (a.Title.Contains(key) || a.Code.Contains(key)))
+                .Count(out var total)
+                .OrderBy(true, a => a.OrderIndex)
+                .ToListAsync();
+
+            var result = _mapper.Map<List<GroupListResponse>>(list);
+            return ResponseEntity.Ok(result);
         }
 
-        public Task<IResponseEntity> SoftDeleteBatchAsync(string[] ids)
-        {
-            throw new NotImplementedException();
-        }
+        // 上面为通用方法实现
+        // 下面为特殊方法实现
 
-        public async Task<IResponseEntity> UpdateAsync(GroupEditRequest req)
-        {
-           
-            if (req == null || req?.Id == "")
-            {
-                return ResponseEntity.Error("更新的实体主键丢失");
-            }
-
-            //var entity = await _sysGroupRepository.GetAsync(req.Id);
-            //if (entity == null || entity?.Id == "")
-            //{
-            //    return ResponseEntity.Error("找不到更新的实体！");
-            //}
-            //_mapper.Map(req, entity);
-
-            var entity = _mapper.Map<SysGroupEntity>(req);            
-            await _sysGroupRepository.UpdateAsync(entity);
-            return ResponseEntity.Ok("更新成功");
-        }
     }
 }
